Split agent and runtime modules
